Learn: move assets into learn folder
