+++ src-ts/tools/learn/assets/data/courses.json
