+++ src-ts/tools/learn/assets/data/learning_resource_providers.json
