+++ src-ts/tools/learn/services/learn-provider/providers/my-certification-progress.provider.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useCallback, useEffect, useMemo, useState } from 'react'
 
-import myCertificationsJSON from '../../../../../assets/data/my-certifications.json'
+import myCertificationsJSON from '../../../assets/data/my-certifications.json'
 
 import { LearnMyCertificationProgress } from './learn-my-certification-progress.model'
 import { LearnMyCertification } from './learn-my-certification.model'
