+++ src-ts/tools/learn/services/learn-provider/providers/my-certifications.provider.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useCallback, useEffect, useMemo, useState } from 'react'
 
-import myCertificationsJSON from '../../../../../assets/data/my-certifications.json'
+import myCertificationsJSON from '../../../assets/data/my-certifications.json'
 
 import { LearnMyCertification } from './learn-my-certification.model'
